requirements.txt
openai
python-dotenv
PyAutoGUI
python3-xlib